Added automatic file hash verification for the base ISO image
Code will attempt tp get the valid file hash from the same directory as wherever the ISO is being sourced from, parse it, and compare it to the computer file hash of the base file hash. If the valid checksum cannot be found/parsed, a warning is printed but the code continues on ahead. If the valid hash is found+parsed but doesnt match the actual computed file hash, the user will be asked if they want to continue or redownload the iso (if ISO source is on the internet) / about (if ISO source is local)